backend/requirements.txt
flask
flask-cors
pymongo
python-docx
pdfplumber
requests
docx2txt
thefuzz
python-dotenv
google-generativeai